% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{cvar}
\alias{cvar}
\title{Coefficiente di variazione}
\usage{
cvar(x)
}
\arguments{
\item{x}{vettore}
}
\value{
Valore del coefficiente di variazione
}
\description{
Calcola il coefficiente di variazione.
Di default, considera solo i casi validi.
}
