% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devianza.R
\name{devianza}
\alias{devianza}
\title{Devianza}
\usage{
devianza(x, ...)
}
\arguments{
\item{x}{vettore}

\item{...}{altri argomenti ereditati da \code{\link[stats]{cor}}}
}
\value{
Valore della devianza
}
\description{
Calcola la devianza come \code{\link[stats]{cor}} * \code{\link[LabRS]{nval}} - 1.
Sono considerati i soli casi validi.
}
