% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{diffmm}
\alias{diffmm}
\title{Coefficiente di asimmetria di Pearson}
\usage{
diffmm(x)
}
\arguments{
\item{x}{vettore}
}
\value{
Valore del coefficiente di asimmetria
}
\description{
Calcola il coefficiente di asimmetria di Pearson, basato sulla differenza
fra media e mediana (i casi mancanti sono eliminati dal computo):
3 * ( mean - median ) / sd.
}
\examples{
diffmm(cars$speed)
}
