% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{esm}
\alias{esm}
\title{Errore standard (della media)}
\usage{
esm(x, vt = 1.96)
}
\arguments{
\item{x}{vettore}

\item{vt}{qualunque valore della distribuzione t di Student:
default = 1.96 (int. conf 95\%); per un intervallo di confidenza
del 99\% impostarlo a 2.58.}
}
\value{
Valore dell'errore standard, media, estremi dell'intervallo di confidenza, errore standard relativo (\%).
}
\description{
Calcola l'errore standard dalla media, e gli estremi dell'intervallo di confidenza.
Di default, considera solo i casi validi e l'intervallo di confidenza è del 95\%.
}
\examples{
esm(cars$speed)
}
