% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expchart.R
\name{expchart}
\alias{expchart}
\title{Esporta i grafici}
\usage{
expchart(device = png, file, ppi = 72, hprop = 6.7, wprop = 6.7, ...)
}
\arguments{
\item{device}{default = png; formato immagine (png, jpeg, tiff, bmp; non pdf e svg) vedi \code{\link[grDevices]{dev.copy}}}

\item{file}{nome del file di esportazione (con estensione in base al tipo di
immagine - device)}

\item{ppi}{default = 72; risoluzione dell'immagine}

\item{hprop}{default = 6.7; altezza dell'immagine, in proporzione ai ppi (es: 4)}

\item{wprop}{default = 6.7; larghezza dell'immagine, in proporzione ai ppi (es: 4)}

\item{...}{altri argomenti ereditati \code{\link[grDevices]{dev.copy}}}
}
\value{
file immagine
}
\description{
Esportazione semplificata di immagini non vettoriali dei grafici con \code{\link[grDevices]{dev.copy}}.
I valori di default sono impostati in modo da esportare una immagine png con
la risoluzione e le dimensioni prodotte da R di default.
}
\examples{
\dontrun{
expchart(png, file = "chart.png", ppi = 300, hprop = 4, wprop = 4)
}
}
