% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdata.R
\name{expdata}
\alias{expdata}
\title{Esporta un dataset in csv (standard italiano)}
\usage{
expdata(x, file, ...)
}
\arguments{
\item{x}{dataset}

\item{file}{nome del file di esportazione (con estensione csv)}

\item{...}{argomenti ereditati da \code{\link[utils]{write.table}}}
}
\value{
file in formato csv
}
\description{
Esportazione semplificata di un dataframe in formato csv,
con \code{\link[utils]{write.table}}, e
secondo lo standard italiano:
sep = ";",
dec = "," (separatore di decimali = virgola), NA = "" (celle vuote),
row.names = FALSE
(il dataframe da esportare ha una colonna / variabile per gli
identificativi di caso).
}
\examples{
data("MYSLID")
expdata(MYSLID, file = "MYSLID.csv")
}
