% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{exptbl}
\alias{exptbl}
\title{Esporta una tabella in txt tab delimited}
\usage{
exptbl(x, file, ...)
}
\arguments{
\item{x}{oggetto table o xtabs, matrice}

\item{file}{nome del file di esportazione (con estensione txt)}

\item{...}{altri argomenti ereditati da \code{\link[utils]{write.table}}}
}
\value{
file in formato txt
}
\description{
Esportazione semplificata di una tabella in formato txt tab delimited,
secondo lo standard Excel in lingua italiana:
separatore di decimali = virgola, NA = celle vuote,
col.names = NA.
}
