% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{exptbls}
\alias{exptbls}
\title{Esporta una lista di tabelle in un file txt tab delimited}
\usage{
exptbls(x, file, ...)
}
\arguments{
\item{x}{vettore carattere, o lista, con i nomi delle tabelle}

\item{file}{nome del file di esportazione (con estensione txt)}

\item{...}{altri argomenti ereditati \code{\link[utils]{write.table}}}
}
\value{
file in formato txt
}
\description{
Esportazione semplificata di più tabelle in formato txt tab delimited,
secondo lo standard Excel in lingua italiana, in uno stesso file:
separatore di decimali = virgola, NA = celle vuote,
col.names = NA.
}
\examples{
# tutte le tabelle sono esportate nel file tabelle.txt
exptbls(c("tab3", "tab6", "tab8"), file = "tabelle.txt")
exptbls(list("tab2", "tab4", "tab5"), file = "tabelle.txt")
exptbls("tab1", file = "tabelle.txt")
}
