\name{faccende}
\alias{faccende}
\docType{data}
\title{
Faccende domestiche (dati)}
\description{Tabella di contingenza, tratta da \link[factoextra]{housetasks}. Tradotto in italiano}
\usage{data("faccende")}
\format{
  Tabella di contingenza con 13 righe e 4 colonne:
  \describe{
    \item{\code{Moglie}}{vettore numerico}
    \item{\code{Alternati}}{vettore numerico}
    \item{\code{Marito}}{vettore numerico}
    \item{\code{Insieme}}{vettore numerico}
  }
}

\source{
dataset tratto da factoextra, e prima da FactoMineR
}


\keyword{datasets}
