% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequenze.R
\name{frequenze}
\alias{frequenze}
\title{Tabelle di frequenza di un dataframe (formato csv)}
\usage{
frequenze(x, write = TRUE, file, aggiungi = FALSE, ...)
}
\arguments{
\item{x}{dataframe}

\item{write}{logico; esporta o no le frequenze in un file csv}

\item{file}{(se write = TRUE) carattere; nome del file di esportazione (con estensione csv)}

\item{aggiungi}{(se write = TRUE) logico; aggiunge le tabelle ad un file esistente (come append)}

\item{...}{argomenti ereditati da \code{\link[LabRS]{tabfreq}}}
}
\value{
file in formato csv con le tabelle,  o lista con le tabelle
}
\description{
Produce le tabelle delle frequenze per tutte le variabili categoriali
(fattori) di un file, con \code{\link[LabRS]{tabfreq}},
e le esporta in formato csv con \code{\link[LabRS]{exptab}}.
}
\examples{
data("MYSLID")
frequenze(MYSLID, "frequenze.csv")
}
