% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{iqr}
\alias{iqr}
\title{Differenza interquartile}
\usage{
iqr(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{vettore o fattore ordinato}

\item{na.rm}{valore logico (TRUE o FALSE) (vedi \code{\link[stats]{IQR}})}

\item{...}{altri argomenti ereditati da \code{\link[stats]{IQR}}}
}
\value{
quantili (valori o livelli)
}
\description{
Calcola la differenza interquartile anche per i fattori ordinati.
Per default, considera solo i casi validi.
}
\examples{
iqr(Orange$Tree) # fattore ordinato
iqr(cars$speed) # vettore numerico
}
