% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{kabbit}
\alias{kabbit}
\title{Tabelle 'kable' con formato dei numeri in italiano}
\usage{
kabbit(x, ...)
}
\arguments{
\item{x}{oggetto table o xtabs, matrice, o dataframe}

\item{...}{altri argomenti ereditati da \code{\link[knitr]{kable}}}
}
\value{
vettore carattere del codice della tabella
}
\description{
Genera tabelle con standard dei numeri in lingua italiana, con la funzione
kable di knitr:
separatore di decimali = virgola, separatore di migliaia = punto
}
