% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\name{med}
\alias{med}
\title{Mediana (fattori ordinati)}
\usage{
med(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{vettore o fattore ordinato}

\item{na.rm}{valore logico (TRUE o FALSE) (vedi \code{\link[stats]{median}})}

\item{...}{altri argomenti ereditati da \code{\link[stats]{median}}}
}
\value{
mediana (valore o livello)
}
\description{
Calcola la mediana anche per i fattori ordinati. Tratta dalla funzione \code{\link[labstatR]{Me}} 
del pacchetto labstatR.
Per default, considera solo i casi validi.
}
\examples{
med(Orange$Tree) # fattore ordinato
med(cars$speed) # vettore numerico
}
\references{
Stefano M.Iacus and Guido Masarotto (2018). labstatR: Libreria
Del Laboratorio Di Statistica Con R. R package version 1.0.9.
}
