% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{nmiss}
\alias{nmiss}
\title{Casi mancanti}
\usage{
nmiss(x)
}
\arguments{
\item{x}{oggetto (vettore, matrice, dataframe ...)}
}
\value{
Numero di casi mancanti
}
\description{
Numero di casi mancanti in un oggetto.
}
\examples{
nmiss(cars)
nmiss(ChickWeight$Diet)
}
