% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmiss_nval.R
\name{nval}
\alias{nval}
\title{Casi validi}
\usage{
nval(x)
}
\arguments{
\item{x}{oggetto (vettore, matrice, dataframe ...)}
}
\value{
Numero di casi validi
}
\description{
Numero di casi validi in un oggetto.
}
\examples{
nval(cars)
nval(ChickWeight$Diet)
}
