% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Percentuali}
\usage{
percent(x, digits = 8, margin = NULL)
}
\arguments{
\item{x}{array (dataframe, numeric).}

\item{digits}{numero dei decimali (vedi \code{\link[base]{round}})}

\item{margin}{marginale (vedi \code{\link[base]{prop.table}})}
}
\value{
matrice dei valori percentuali
}
\description{
Calcola le percentuali, con arrotondamento (se x = dataframe
-> as.matrix, se x = double -> as.array).
}
\examples{
percent(table(ChickWeight$Diet), 2)
percent(HairEyeColor, 1, margin = c(3,2))
}
