% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{quant}
\alias{quant}
\title{Quantili}
\usage{
quant(x, probs = c(0.25, 0.5, 0.75), na.rm = TRUE, ...)
}
\arguments{
\item{x}{vettore o fattore ordinato}

\item{probs}{vettore numerico di probabilità con valori compresi fra 0 e 1
(vedi \code{\link[stats]{quantile}})}

\item{na.rm}{valore logico (TRUE o FALSE) (vedi \code{\link[stats]{quantile}})}

\item{...}{altri argomenti ereditati da \code{\link[stats]{quantile}}}
}
\value{
quantili (valori o livelli)
}
\description{
Calcola i quantili anche per i fattori ordinati.
Per default, restituisce i soli quartili e considera solo i casi validi.
}
\examples{
quant(Orange$Tree) # fattore ordinato
quant(cars$speed) # vettore numerico
}
