% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabcont.R
\name{tabcont}
\alias{tabcont}
\title{Tabelle di contingenza}
\usage{
tabcont(x, y, digits = 1, totali = TRUE)
}
\arguments{
\item{x}{fattore: variabile di riga}

\item{y}{fattore: variabile di colonna}

\item{digits}{numero dei decimali (vedi \code{\link[base]{round}})}

\item{totali}{logico: TRUE = con totali di colonna in VA}
}
\value{
una matrice con la tabella
}
\description{
Produce una tabella di contingenza per la presentazione dei dati,
in valori percentuali di colonna (un decimale di default), totali
di colonna in VA opzionali
}
\examples{
data("MYSLID")
tabcont(MYSLID$Lingua, MYSLID$Genere)
tabcont(MYSLID$Lingua, MYSLID$Genere, totali = FALSE)
}
