% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{tabfreq}
\alias{tabfreq}
\title{Tabelle di frequenza}
\usage{
tabfreq(x, digits = 2, totali = TRUE)
}
\arguments{
\item{x}{fattore}

\item{digits}{numero dei decimali (vedi \code{\link[base]{round}})}

\item{totali}{logico: TRUE = con totale e casi mancanti}
}
\value{
una matrice con due colonne
}
\description{
Produce una tabella delle frequenze per la presentazione dei dati, 
in valori assoluti e percentuali (due decimali di default)
}
\examples{
tabfreq(ChickWeight$Diet)
tabfreq(ChickWeight$Diet, totali = FALSE)
}
