% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esm.R
\name{esm}
\alias{esm}
\title{Errore standard (della media)}
\usage{
esm(x, q = 1.96, digits = 6)
}
\arguments{
\item{x}{vettore}

\item{q}{quantile (o qualunque valore) della distribuzione normale
default = 1.96 (int. conf 95\%); per un intervallo di confidenza
del 99\% impostarlo a 2.58.}

\item{digits}{decimali}
}
\value{
Valore della media, della deviazione standard, dell'errore standard,
estremi dell'intervallo di confidenza, valore p dell'area.
}
\description{
Calcola l'errore standard dalla media, e gli estremi dell'intervallo di confidenza.
Di default, considera solo i casi validi con un intervallo di confidenza del 95\%.
}
\examples{
esm(cars$speed)
}
