% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{expdata}
\alias{expdata}
\title{Esporta un dataset in txt tab delimited}
\usage{
expdata(x, file, ...)
}
\arguments{
\item{x}{dataset}

\item{file}{nome del file di esportazione (con estensione txt)}

\item{...}{altri argomenti ereditati da \code{\link[utils]{write.table}}}
}
\value{
file in formato txt
}
\description{
Esportazione semplificata di un dataset in formato txt tab delimited,
secondo lo standard Excel in lingua italiana:
separatore di decimali = virgola, NA = celle vuote. L'attributo
row.names = FALSE di default (il dataset da esportare ha una colonna
per gli ID), ma può essere modificato.
}
