% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{med}
\alias{med}
\title{Mediana}
\usage{
med(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{vettore o fattore ordinato}

\item{na.rm}{valore logico (TRUE o FALSE) (vedi \code{\link[stats]{median}})}

\item{...}{altri argomenti ereditati da \code{\link[stats]{median}}}
}
\value{
mediana (valore o livello)
}
\description{
Calcola la mediana anche per i fattori ordinati.
Per default, considera solo i casi validi.
}
\examples{
med(Orange$Tree) # fattore ordinato
med(cars$speed) # vettore numerico
}
