% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{percent}
\alias{percent}
\title{Percentuali}
\usage{
percent(x, digits = 8, margin = NULL)
}
\arguments{
\item{x}{vettore, matrice, tabella}

\item{digits}{numero dei decimali (vedi \code{\link[base]{round}})}

\item{margin}{marginale (vedi \code{\link[base]{prop.table}})}
}
\value{
valori percentuali
}
\description{
Calcola le percentuali, con arrotondamento.
}
\examples{
percent(table(ChickWeight$Diet), 2)
percent(HairEyeColor, 1, margin = c(3,2))
}
