\name{tabM}
\alias{tabM}
\docType{data}
\title{
Tabella "M" (dati)
}
\description{
Tabella di contingenza (tipo "matrice"), adattata da un esempio del pacchetto \link[stats]{stats}.
}

\usage{data("tabM")}

\format{
  Frequenze condizionate delle variabili:
  \describe{
    \item{\code{x}}{identificazione di partito: vettore numerico}
    \item{\code{y}}{genere: vettore numerico}
  }
}

\source{Tabella tratta da \link[stats]{chisq.test}}

\references{
Agresti, Alan. 2007. An introduction to categorical data analysis. 2nd ed. Hoboken, NJ: Wiley-Interscience.
}

\keyword{datasets}
